Add batch helper
